clean folders structure
