+++ src/views/Hostels/index.tsx
@@ -1,0 +1,11 @@
+import React, { useContext } from 'react';
+
+function Hostels(): JSX.Element {
+  return (
+    <div className="Hostels">
+      <h1>Hostel component work</h1>
+    </div>
+  );
+}
+
+export default Hostels;
