+++ src/views/UserList/index.tsx
@@ -1,0 +1,11 @@
+import React, { useContext } from 'react';
+
+function UserList(): JSX.Element {
+  return (
+    <div className="UserList">
+      <h1>UserList component work</h1>
+    </div>
+  );
+}
+
+export default UserList;
