+++ src/views/Vehicles/index.tsx
@@ -1,0 +1,11 @@
+import React, { useContext } from 'react';
+
+function Vehicles(): JSX.Element {
+  return (
+    <div className="Vehicles">
+      <h1>Component Vehicles works</h1>
+    </div>
+  );
+}
+
+export default Vehicles;
