--- src/views/hostels/hostels.tsx
@@ -1,20 +1,0 @@
-import React, { useContext } from 'react';
-// import { MainStore } from '../../store/MainStore';
-// import { UserStore } from '../../store/UserStore';
-
-
-function Hostels(): JSX.Element {
-  // const { state } = useContext(MainStore);
-  // const { userState } = useContext(UserStore);
-
-  return (
-    <div className="Hostels">
-      <h1>
-        Hostel component work
-      </h1>
-    </div>
-  );
-}
-
-export default Hostels;
-
