--- src/views/userList/userList.tsx
@@ -1,20 +1,0 @@
-import React, { useContext } from 'react';
-// import { MainStore } from '../../store/MainStore';
-// import { UserStore } from '../../store/UserStore';
-
-
-function UserList(): JSX.Element {
-  // const { state } = useContext(MainStore);
-  // const { userState } = useContext(UserStore);
-
-  return (
-    <div className="UserList">
-      <h1>
-        UserList component work
-      </h1>
-    </div>
-  );
-}
-
-export default UserList;
-
