--- src/views/userOverview/userOverview.sass
@@ -1,3 +1,0 @@
-.UserOverview {
-  
-}
