+++ src/views/userOverview/userOverview.tsx
@@ -1,12 +1,6 @@
 import React, { useContext } from 'react';
-// import { MainStore } from '../../store/MainStore';
-// import { UserStore } from '../../store/UserStore';
-
 
 function UserOverview(): JSX.Element {
-  // const { state } = useContext(MainStore);
-  // const { userState } = useContext(UserStore);
-
   return (
     <div className="UserOverview">
       <h1>Component UserOverview works</h1>
@@ -15,4 +9,3 @@ function UserOverview(): JSX.Element {
 }
 
 export default UserOverview;
-
