--- src/views/vehicles/vehicles.tsx
@@ -1,18 +1,0 @@
-import React, { useContext } from 'react';
-// import { MainStore } from '../../store/MainStore';
-// import { UserStore } from '../../store/UserStore';
-
-
-function Vehicles(): JSX.Element {
-  // const { state } = useContext(MainStore);
-  // const { userState } = useContext(UserStore);
-
-  return (
-    <div className="Vehicles">
-      <h1>Component Vehicles works</h1>
-    </div>
-  );
-}
-
-export default Vehicles;
-
